refactor: rename scryfall-api files to kebab-case
getCardName.mjs -> get-card-name.mjs, searchName.mjs -> search-name.mjs,
matching the kebab-case filename convention used everywhere else in the
repo (match-name.mjs, card-hash-cache.mjs, needs-attention-store.mjs, ...).
